PixelFlipper

Documentation for PixelFlipper.

PixelFlipper.AbstractImputer

PixelFlipper.AbstractSelector

PixelFlipper.ConstantImputer

PixelFlipper.PixelFlipping

PixelFlipper.PixelSelector

PixelFlipper.evaluate

PixelFlipper.select

PixelFlipper.AbstractImputer — Type

AbstractImputer

Abstract supertype of all imputers. Given an input and indices from a selector, this imputes values of an array at the given index.

PixelFlipper.AbstractSelector — Type

AbstractSelector

Abstract supertype of all selectors. Given an Explanation or WHCN array, all AbstractSelector return an iterator of values to be imputed.

PixelFlipper.ConstantImputer — Type

ConstantImputer(value)

Imputes a constant value into a given array. Defaults to 0.0f0.

PixelFlipper.PixelFlipping — Type

PixelFlipping([; selector, imputer, steps])

Computes pixel flipping curves.

Keyword arguments

selector::AbstractSelector: Specify input selector. Defaults to PixelSelector(:norm).

imputer::AbstractImputer: Specify input imputer. Defaults to ConstantImputer{Float32}(0.0f0) of value zero.

steps::Int: Specify number of imputation steps. Has to be smaller than the amount of selectable inputs in a sample. Defaults to 25.

show_progress:Bool: Show progress meter while sampling augmentations. Defaults to true.

device: Select array type to run pixel flipping on, e.g. CUDA.jl's cu for GPU support. Defaults to CPU Array.

PixelFlipper.PixelSelector — Type

PixelSelector()
PixelSelector(; reduce=:norm)

Reduces color channels in an Explanation according to reduce and returns an iterator over the indices of sorted values.

Keyword arguments

reduce::Symbol: Selects how color channels are reduced to a single number to apply a color scheme. The following methods can be selected, which are then applied over the color channels for each "pixel" in the array:

:sum: sum up color channels

:norm: compute 2-norm over the color channels

:maxabs: compute maximum(abs, x) over the color channels

:sumabs: compute sum(abs, x) over the color channels

:abssum: compute abs(sum(x)) over the color channels

Defaults to :norm.

PixelFlipper.evaluate — Method

evaluate(pixelflipping, model, input, explanation::Explanation)
evaluate(pixelflipping, model, input, values::AbstractArray) 
evaluate(pixelflipping, model, input, analyzer::AbstractXAIMethod)

Run the PixelFlipping method on the given model, input and explanation.

PixelFlipper.select — Method

select(x, selector)

Return matrix of CartesianIndices of x sorted by decreasing value. Requires x to be in WHCN format, as each column in the output corresponds to an inputs in the batch.

Example

julia> selector = PixelSelector()
PixelSelector(:norm)

julia> A = randn(1, 2, 2, 2)
1×2×2×2 Array{Float64, 4}:
[:, :, 1, 1] =
 -1.9275  -3.01383

[:, :, 2, 1] =
 -0.424713  1.24167

[:, :, 1, 2] =
 -1.36198  1.21235

[:, :, 2, 2] =
 -1.75508  -0.700117

julia> PixelFlipper.select(A, selector)
2×4 Matrix{CartesianIndex{4}}:
 CartesianIndex(1, 2, 1, 1)  CartesianIndex(1, 2, 2, 1)  CartesianIndex(1, 1, 1, 2)  CartesianIndex(1, 1, 2, 2)
 CartesianIndex(1, 1, 1, 1)  CartesianIndex(1, 1, 2, 1)  CartesianIndex(1, 2, 1, 2)  CartesianIndex(1, 2, 2, 2)